- fix minor compiler warnings.

+++ QRCodeGenLib/src/QRCodeGen/QlpQrSegment.pas
@@ -422,7 +422,7 @@ begin
   for LIdx := System.Low(AData) to System.High(AData) do
   begin
     LBits[TBits.Asr32(LIdx, 2)] := LBits[TBits.Asr32(LIdx, 2)] or
-      ((AData[LIdx] and $FF) shl ((not LIdx) shl 3));
+      Int64((AData[LIdx] and $FF) shl ((not LIdx) shl 3));
   end;
   result := TQrSegment.Create(TQrSegmentMode.qsmByte, System.Length(AData),
     LBits, System.Length(AData) * 8);
